pyproject.toml
[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "hug"
src_paths = ["tarraz"]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
default_section = "THIRDPARTY"
known_first_party = "tarraz"

[tool.poetry]
name = "tarraz"
version = "1.2.0"
description = "A cross-stitch image generator. Generates a cross stitch pattern given by a user and generates a DMC colored pattern."
authors = ["Ahmed Jazzar <ahmed@nitfe.com>"]
maintainers = []

license = "GNU Affero General Public License v3"
readme = "README.md"
packages = [{ include = "tarraz" }]
homepage = "https://www.nitfe.com/"
repository = "https://github.com/nitfe/tarraz.git"
documentation = "https://github.com/nitfe/tarraz/blob/main/README.md"
keywords = ["cross-stitch", "pattern", "stitch", "DMC"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.urls]
Issues = "https://github.com/nitfe/tarraz/issues"
Changelog = "https://github.com/nitfe/tarraz/releases"

# Requirements
[tool.poetry.dependencies]
python = ">=3.8"
pillow = ">=9.5.0"

[tool.poetry.dev-dependencies]
pre-commit = { version = "*" }

# Entry point
[tool.poetry.plugins.console_scripts]
"tarraz" = "tarraz.main:main"
